+++ ameriflux_pipeline/__init__.py
@@ -6,7 +6,8 @@
 
 from ameriflux_pipeline.config import Config
 import ameriflux_pipeline.enveditor
-import ameriflux_pipeline.data_validation
+from ameriflux_pipeline.utils.data_validation import DataValidation
+from ameriflux_pipeline.utils.input_validation import InputValidation
 import ameriflux_pipeline.met_data_merge
 import ameriflux_pipeline.pre_pyfluxpro
 import ameriflux_pipeline.post_pyfluxpro