+++ ameriflux_pipeline/utils/data_util.py
@@ -68,5 +68,4 @@ def get_directory(file_path):
     """
     path = pathlib.Path(file_path)
     dir = os.path.dirname(path)
-
     return dir